Don't run GLX_OML_sync_control tests concurrently.
This fixes the swapbuffersmsc-return (with default swap_interval) test,
by not passing it arguments that it doesn't understand.

These tests all require a GLXWindow, so they can't be used in -fbo mode
(and they didn't understand that option anyway).

And though they need the behavior of -auto, none of them parse that
option. These tests have to deal with it in other ways, as illustrated
by commit 1d618f2abbc9f9665505b7082690eebc3bc01a72.

Finally, I think it's a bad idea to run these tests concurrently with
anything else, as interference from other clients could change the
result of attempting to sync to vertical retrace, which is a
screen-global resource.

Signed-off-by: Jamey Sharp <jamey@minilop.net>
Signed-off-by: Theo Hill <Theo0x48@gmail.com>